+++ test/core/test_scoring.py
@@ -6,6 +6,6 @@ def test_download_template():
 
 
 def test_download_scoring():
-    scoring(['dummy',  '--challenge', 'D3C1', '--download-goldstandard'])
-    scoring(['dummy',  '--challenge', 'D3C2', '--download-goldstandard', '--sub-challenge', 'cytokine'])
+    scoring(['dummy',  '--challenge', 'D3C1', '--download-template'])
+    scoring(['dummy',  '--challenge', 'D3C2', '--download-template', '--sub-challenge', 'cytokine'])