+++ doc/source/changelog.rst
@@ -5,6 +5,7 @@ ChangeLog
 ----------
 
 
+ * NEWS: some new classes :class:`dreamtools.dream2` related to DREAM2
  * NEWS: add :class:`dreamtools.dream5.scoring.D5C1` challenge in Dreamtools
  * NEWS: add :class:`dreamtools.dream3.scoring.D3C2` challenge in Dreamtools
  * NEWS: add :class:`dreamtools.dream3.scoring.D3C3` challenge in Dreamtools